Avoid emulating wider atomics
Generally, Rust eschews emulating atomics where it can.
This is done precisely so people can write interrupt handlers in Rust,
without accidentally arriving in other interrupt handlers,
or other complications that such things tend to induce.

No solution is yet proposed for the xtensa-esp32s2-espidf target.
However, it would be less-than-ideal if an ecosystem grew up around
Rust-on-Xtensa that was dependent on something historically rejected by
upstream Rust in the cases where it was easily avoidable.